Flatter module hierarchy
