+++ lib/rpush/client/active_record/apns2/notification.rb
@@ -2,10 +2,10 @@ module Rpush
   module Client
     module ActiveRecord
       module Apns2
-        class Notification < Rpush::Client::ActiveRecord::Apns::Notification
-          def max_payload_bytesize
-            Rpush::Client::ActiveModel::Apns2::MAX_PAYLOAD_BYTESIZE
-          end
+        class Notification < Rpush::Client::ActiveRecord::Notification
+          include Rpush::Client::ActiveModel::Apns::Notification
+          include Rpush::Client::ActiveModel::Apns2::Notification
+          include Rpush::Client::ActiveRecord::Apns::ActiveRecordSerializableNotification
         end
       end
     end
